added implicitly declared header, fixed wrong datatype declaration

+++ libavformat/concat.c
@@ -36,7 +36,7 @@ extern int ff_concatgen_read_packet(AVFormatContext *s, AVPacket *pkt);
 
 extern int ff_concatgen_read_seek(AVFormatContext *s, int stream_index, int64_t pts, int flags);
 
-extern int ff_concatgen_read_timestamp(AVFormatContext *s, int stream_index, int64_t *pos, int64_t pos_limit);
+extern int64_t ff_concatgen_read_timestamp(AVFormatContext *s, int stream_index, int64_t *pos, int64_t pos_limit);
 
 extern int ff_concatgen_read_close(AVFormatContext *s);
